Restructure Patient requests
