+++ app/Http/Controllers/PatientController.php
@@ -4,9 +4,7 @@ namespace App\Http\Controllers;
 
 use Validator;
 use App\Patient;
-use App\Http\Requests\ShowPatient;
-use App\Http\Requests\StorePatient;
-use Illuminate\Database\Eloquent\ModelNotFoundException;
+use App\Http\Requests\Patient as PatientRequest;
 
 class PatientController extends Controller
 {
@@ -21,7 +19,7 @@ class PatientController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(StorePatient $request)
+    public function store(PatientRequest\Store $request)
     {
         $patient = new Patient([
             'first_name' => $request->input('firstName'),
@@ -59,7 +57,7 @@ class PatientController extends Controller
      * @param  \App\patient  $patient
      * @return \Illuminate\Http\Response
      */
-    public function show(Patient $patient, ShowPatient $request)
+    public function show(PatientRequest\Show $request, Patient $patient)
     {
         $gotPatient = [
             'id' => $patient->id,
