+++ app/Http/Requests/Patient/Store.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Patient;
 
 use App\Patient;
-use Illuminate\Foundation\Http\FormRequest;
+use App\Http\Requests\BaseRequest;
 
-class StorePatient extends BaseRequest
+class Store extends BaseRequest
 {
     /**
      * Get the validation rules that apply to the request.
